chore: hidde helper classes
